+++ src/main/java/com/yeoljeong/tripmate/product/infrastructure/messaging/PaymentRefundedEventListener.java
@@ -1,8 +1,8 @@
-package com.yeoljeong.tripmate.product.infrastructure.messaging.consumer;
+package com.yeoljeong.tripmate.product.infrastructure.messaging;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.yeoljeong.tripmate.event.PlanUnitParticipantAddedEvent;
-import com.yeoljeong.tripmate.event.enums.PlanTopic;
+import com.yeoljeong.tripmate.event.PaymentRefundedEvent;
+import com.yeoljeong.tripmate.event.enums.PaymentTopic;
 import com.yeoljeong.tripmate.exception.BusinessException;
 import com.yeoljeong.tripmate.product.application.service.command.ProductScheduleCommandService;
 import lombok.RequiredArgsConstructor;
@@ -14,33 +14,31 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class PlanParticipantAddedEventListener {
+public class PaymentRefundedEventListener {
 
   private final ProductScheduleCommandService productScheduleCommandService;
   private final ObjectMapper objectMapper;
 
   @KafkaListener(
-      topics = PlanTopic.PLAN_UNIT_PARTICIPANT_ADDED_TOPIC,
+      topics = PaymentTopic.PAYMENT_REFUNDED_TOPIC,
       groupId = "company-service"
   )
-  public void handleScheduleParticipantAdded(
+  public void handlePaymentRefunded(
       String message,
       Acknowledgment ack
   ) {
     try {
       // 직접 역직렬화
-      PlanUnitParticipantAddedEvent event =
-          objectMapper.readValue(message, PlanUnitParticipantAddedEvent.class);
+      PaymentRefundedEvent event =
+          objectMapper.readValue(message, PaymentRefundedEvent.class);
 
-      log.info("plan.participant.added 이벤트 수신 - scheduleId={}, quantity={}",
+      log.info("payment.refunded 이벤트 수신 - scheduleId={}, quantity={}",
           event.scheduleId(), event.quantity());
 
-      // 재고 차감 서비스 호출
-      productScheduleCommandService.deductStock(
+      // 재고 추가 서비스 호출
+      productScheduleCommandService.increaseStock(
           event.productId(),
           event.scheduleId(),
-          event.planUnitId(),
-          event.userId(),
           event.quantity()
       );
 
@@ -51,11 +49,11 @@ public class PlanParticipantAddedEventListener {
 
     } catch (BusinessException e) {
       // 비즈니스 예외는 재시도 불필요 → ack하고 넘어감
-      log.warn("[PlanParticipantAddedListener] 비즈니스 예외 - error: {}", e.getMessage());
+      log.warn("[PaymentRefundedListener] 비즈니스 예외 - error: {}", e.getMessage());
       ack.acknowledge();
     } catch (Exception e) {
       // 시스템 예외는 재시도
-      log.error("[PlanParticipantAddedListener] 처리 실패", e);
+      log.error("[PaymentRefundedListener] 처리 실패", e);
       throw new RuntimeException(e);
     }
   }
