+++ src/main/java/com/yeoljeong/tripmate/product/infrastructure/persistence/jpa/ProductOutboxJpaRepository.java
@@ -4,7 +4,6 @@ import com.yeoljeong.tripmate.domain.constants.OutboxStatus;
 import com.yeoljeong.tripmate.product.infrastructure.outbox.ProductOutbox;
 import java.util.List;
 import java.util.UUID;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductOutboxJpaRepository extends JpaRepository<ProductOutbox, UUID> {
